# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Magic Key Property enquiry hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EnquiryResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type BtlEnquiryInputIsFirstTimeLandlord = typeof BtlEnquiryInputIsFirstTimeLandlord[keyof typeof BtlEnquiryInputIsFirstTimeLandlord];


export const BtlEnquiryInputIsFirstTimeLandlord = {
  yes: 'yes',
  no: 'no',
} as const;

export type BtlEnquiryInputInvestmentGoal = typeof BtlEnquiryInputInvestmentGoal[keyof typeof BtlEnquiryInputInvestmentGoal];


export const BtlEnquiryInputInvestmentGoal = {
  capital_growth: 'capital_growth',
  rental_income: 'rental_income',
  portfolio_building: 'portfolio_building',
  tax_planning: 'tax_planning',
  other: 'other',
} as const;

export type BtlEnquiryInputOwnershipPreference = typeof BtlEnquiryInputOwnershipPreference[keyof typeof BtlEnquiryInputOwnershipPreference];


export const BtlEnquiryInputOwnershipPreference = {
  personal_name: 'personal_name',
  limited_company: 'limited_company',
  unsure: 'unsure',
} as const;

export type BtlEnquiryInputPropertyType = typeof BtlEnquiryInputPropertyType[keyof typeof BtlEnquiryInputPropertyType];


export const BtlEnquiryInputPropertyType = {
  single_let: 'single_let',
  hmo: 'hmo',
  serviced_accommodation: 'serviced_accommodation',
  multi_unit: 'multi_unit',
  commercial_conversion: 'commercial_conversion',
  unsure: 'unsure',
} as const;

export type BtlEnquiryInputFinanceRequired = typeof BtlEnquiryInputFinanceRequired[keyof typeof BtlEnquiryInputFinanceRequired];


export const BtlEnquiryInputFinanceRequired = {
  yes: 'yes',
  no: 'no',
  unsure: 'unsure',
} as const;

export interface BtlEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  currentPostcode: string;
  isFirstTimeLandlord: BtlEnquiryInputIsFirstTimeLandlord;
  investmentGoal: BtlEnquiryInputInvestmentGoal;
  /** @nullable */
  preferredInvestmentLocation?: string | null;
  /** @nullable */
  budgetDepositAvailable?: string | null;
  /** @nullable */
  expectedPurchasePriceRange?: string | null;
  ownershipPreference: BtlEnquiryInputOwnershipPreference;
  /** @nullable */
  existingPropertiesOwned?: string | null;
  propertyType: BtlEnquiryInputPropertyType;
  financeRequired: BtlEnquiryInputFinanceRequired;
  /** @nullable */
  expectedRentalIncome?: string | null;
  /** @nullable */
  timescaleToInvest?: string | null;
  /** @nullable */
  questionsOrConcerns?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}

export type MortgageEnquiryInputMortgageType = typeof MortgageEnquiryInputMortgageType[keyof typeof MortgageEnquiryInputMortgageType];


export const MortgageEnquiryInputMortgageType = {
  residential: 'residential',
  buy_to_let: 'buy_to_let',
  remortgage: 'remortgage',
  product_transfer: 'product_transfer',
  limited_company_btl: 'limited_company_btl',
  bridging_development: 'bridging_development',
} as const;

export type MortgageEnquiryInputCreditHistory = typeof MortgageEnquiryInputCreditHistory[keyof typeof MortgageEnquiryInputCreditHistory];


export const MortgageEnquiryInputCreditHistory = {
  excellent: 'excellent',
  good: 'good',
  fair: 'fair',
  poor: 'poor',
  unsure: 'unsure',
} as const;

/**
 * @nullable
 */
export type MortgageEnquiryInputHasMortgageOffer = typeof MortgageEnquiryInputHasMortgageOffer[keyof typeof MortgageEnquiryInputHasMortgageOffer] | null;


export const MortgageEnquiryInputHasMortgageOffer = {
  yes: 'yes',
  no: 'no',
} as const;

export interface MortgageEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  postcode: string;
  mortgageType: MortgageEnquiryInputMortgageType;
  /** @nullable */
  purchasePriceOrValue?: string | null;
  /** @nullable */
  mortgageAmountRequired?: string | null;
  /** @nullable */
  depositOrEquityAvailable?: string | null;
  employmentStatus: string;
  /** @nullable */
  annualIncome?: string | null;
  creditHistory: MortgageEnquiryInputCreditHistory;
  /** @nullable */
  propertyLocation?: string | null;
  /** @nullable */
  timescale?: string | null;
  /** @nullable */
  hasMortgageOffer?: MortgageEnquiryInputHasMortgageOffer;
  /** @nullable */
  preferredContactMethod?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}

export type TaxEnquiryInputAlreadyInvestingInProperty = typeof TaxEnquiryInputAlreadyInvestingInProperty[keyof typeof TaxEnquiryInputAlreadyInvestingInProperty];


export const TaxEnquiryInputAlreadyInvestingInProperty = {
  yes: 'yes',
  no: 'no',
} as const;

export type TaxEnquiryInputAlreadyHasLimitedCompany = typeof TaxEnquiryInputAlreadyHasLimitedCompany[keyof typeof TaxEnquiryInputAlreadyHasLimitedCompany];


export const TaxEnquiryInputAlreadyHasLimitedCompany = {
  yes: 'yes',
  no: 'no',
} as const;

export type TaxEnquiryInputServiceRequired = typeof TaxEnquiryInputServiceRequired[keyof typeof TaxEnquiryInputServiceRequired];


export const TaxEnquiryInputServiceRequired = {
  limited_company_setup: 'limited_company_setup',
  bookkeeping: 'bookkeeping',
  annual_accounts: 'annual_accounts',
  corporation_tax: 'corporation_tax',
  self_assessment: 'self_assessment',
  vat: 'vat',
  payroll: 'payroll',
  spv_setup: 'spv_setup',
  property_tax_planning: 'property_tax_planning',
  other: 'other',
} as const;

/**
 * @nullable
 */
export type TaxEnquiryInputNeedCompaniesHouseHelp = typeof TaxEnquiryInputNeedCompaniesHouseHelp[keyof typeof TaxEnquiryInputNeedCompaniesHouseHelp] | null;


export const TaxEnquiryInputNeedCompaniesHouseHelp = {
  yes: 'yes',
  no: 'no',
} as const;

/**
 * @nullable
 */
export type TaxEnquiryInputNeedBusinessBankAccountHelp = typeof TaxEnquiryInputNeedBusinessBankAccountHelp[keyof typeof TaxEnquiryInputNeedBusinessBankAccountHelp] | null;


export const TaxEnquiryInputNeedBusinessBankAccountHelp = {
  yes: 'yes',
  no: 'no',
} as const;

/**
 * @nullable
 */
export type TaxEnquiryInputNeedTaxAdviceBeforePurchase = typeof TaxEnquiryInputNeedTaxAdviceBeforePurchase[keyof typeof TaxEnquiryInputNeedTaxAdviceBeforePurchase] | null;


export const TaxEnquiryInputNeedTaxAdviceBeforePurchase = {
  yes: 'yes',
  no: 'no',
} as const;

export interface TaxEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  alreadyInvestingInProperty: TaxEnquiryInputAlreadyInvestingInProperty;
  alreadyHasLimitedCompany: TaxEnquiryInputAlreadyHasLimitedCompany;
  serviceRequired: TaxEnquiryInputServiceRequired;
  /** @nullable */
  numberOfPropertiesOwned?: string | null;
  /** @nullable */
  estimatedAnnualRentalIncome?: string | null;
  /** @nullable */
  propertiesOwnedPersonallyOrCompany?: string | null;
  /** @nullable */
  needCompaniesHouseHelp?: TaxEnquiryInputNeedCompaniesHouseHelp;
  /** @nullable */
  needBusinessBankAccountHelp?: TaxEnquiryInputNeedBusinessBankAccountHelp;
  /** @nullable */
  needTaxAdviceBeforePurchase?: TaxEnquiryInputNeedTaxAdviceBeforePurchase;
  /** @nullable */
  preferredContactMethod?: string | null;
  /** @nullable */
  messageOrEnquiry?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}