# lib/api-zod/src/generated/types/btlEnquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Magic Key Property enquiry hub API
 * OpenAPI spec version: 0.1.0
 */
import type { BtlEnquiryInputFinanceRequired } from './btlEnquiryInputFinanceRequired';
import type { BtlEnquiryInputInvestmentGoal } from './btlEnquiryInputInvestmentGoal';
import type { BtlEnquiryInputIsFirstTimeLandlord } from './btlEnquiryInputIsFirstTimeLandlord';
import type { BtlEnquiryInputOwnershipPreference } from './btlEnquiryInputOwnershipPreference';
import type { BtlEnquiryInputPropertyType } from './btlEnquiryInputPropertyType';

export interface BtlEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  currentPostcode: string;
  isFirstTimeLandlord: BtlEnquiryInputIsFirstTimeLandlord;
  investmentGoal: BtlEnquiryInputInvestmentGoal;
  /** @nullable */
  preferredInvestmentLocation?: string | null;
  /** @nullable */
  budgetDepositAvailable?: string | null;
  /** @nullable */
  expectedPurchasePriceRange?: string | null;
  ownershipPreference: BtlEnquiryInputOwnershipPreference;
  /** @nullable */
  existingPropertiesOwned?: string | null;
  propertyType: BtlEnquiryInputPropertyType;
  financeRequired: BtlEnquiryInputFinanceRequired;
  /** @nullable */
  expectedRentalIncome?: string | null;
  /** @nullable */
  timescaleToInvest?: string | null;
  /** @nullable */
  questionsOrConcerns?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}