# lib/api-zod/src/generated/types/enquiryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Magic Key Property enquiry hub API
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryResult {
  success: boolean;
  message: string;
}