# lib/api-zod/src/generated/types/mortgageEnquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Magic Key Property enquiry hub API
 * OpenAPI spec version: 0.1.0
 */
import type { MortgageEnquiryInputCreditHistory } from './mortgageEnquiryInputCreditHistory';
import type { MortgageEnquiryInputHasMortgageOffer } from './mortgageEnquiryInputHasMortgageOffer';
import type { MortgageEnquiryInputMortgageType } from './mortgageEnquiryInputMortgageType';

export interface MortgageEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  postcode: string;
  mortgageType: MortgageEnquiryInputMortgageType;
  /** @nullable */
  purchasePriceOrValue?: string | null;
  /** @nullable */
  mortgageAmountRequired?: string | null;
  /** @nullable */
  depositOrEquityAvailable?: string | null;
  employmentStatus: string;
  /** @nullable */
  annualIncome?: string | null;
  creditHistory: MortgageEnquiryInputCreditHistory;
  /** @nullable */
  propertyLocation?: string | null;
  /** @nullable */
  timescale?: string | null;
  /** @nullable */
  hasMortgageOffer?: MortgageEnquiryInputHasMortgageOffer;
  /** @nullable */
  preferredContactMethod?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}