# lib/api-zod/src/generated/types/taxEnquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Magic Key Property enquiry hub API
 * OpenAPI spec version: 0.1.0
 */
import type { TaxEnquiryInputAlreadyHasLimitedCompany } from './taxEnquiryInputAlreadyHasLimitedCompany';
import type { TaxEnquiryInputAlreadyInvestingInProperty } from './taxEnquiryInputAlreadyInvestingInProperty';
import type { TaxEnquiryInputNeedBusinessBankAccountHelp } from './taxEnquiryInputNeedBusinessBankAccountHelp';
import type { TaxEnquiryInputNeedCompaniesHouseHelp } from './taxEnquiryInputNeedCompaniesHouseHelp';
import type { TaxEnquiryInputNeedTaxAdviceBeforePurchase } from './taxEnquiryInputNeedTaxAdviceBeforePurchase';
import type { TaxEnquiryInputServiceRequired } from './taxEnquiryInputServiceRequired';

export interface TaxEnquiryInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  mobileNumber: string;
  alreadyInvestingInProperty: TaxEnquiryInputAlreadyInvestingInProperty;
  alreadyHasLimitedCompany: TaxEnquiryInputAlreadyHasLimitedCompany;
  serviceRequired: TaxEnquiryInputServiceRequired;
  /** @nullable */
  numberOfPropertiesOwned?: string | null;
  /** @nullable */
  estimatedAnnualRentalIncome?: string | null;
  /** @nullable */
  propertiesOwnedPersonallyOrCompany?: string | null;
  /** @nullable */
  needCompaniesHouseHelp?: TaxEnquiryInputNeedCompaniesHouseHelp;
  /** @nullable */
  needBusinessBankAccountHelp?: TaxEnquiryInputNeedBusinessBankAccountHelp;
  /** @nullable */
  needTaxAdviceBeforePurchase?: TaxEnquiryInputNeedTaxAdviceBeforePurchase;
  /** @nullable */
  preferredContactMethod?: string | null;
  /** @nullable */
  messageOrEnquiry?: string | null;
  consentToContact: boolean;
  privacyAcknowledged: boolean;
}